Improvement for PR #1 - Title will be taken from the focused item
Improvement for the option of using the addon when kodi is not playing,
instead of typing.
(For movies and episodes)
Also were made some fixes for this.

Note: With some addons they not return all the labels, so when it
doesn't get info, it just remain the same as today (I already sent about
it to their developers)
On library it works always.